requirements.txt
passlib==1.7.0
py-postgresql==1.1.0
PySocks==1.6.1
pytz==2016.7
repoze.lru==0.6
SimpleFramework==1.2.6
SQLAlchemy==1.1.4
suds-py3==1.3.2.0
